requirements.txt
flask
flask-mysql
flask-login
pymysql
flask_sqlalchemy
sqlalchemy
python-decouple
gunicorn
jinja2
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-speech